conditional user rol for delete

+++ src/server/controller/establishmentControllers/establishmentControllers.js
@@ -1,6 +1,7 @@
 const debug = require("debug")("vlcSinGluten:server:controller:establishment");
 const chalk = require("chalk");
 const Establishment = require("../../../database/models/Establishment");
+const { rolAdmin } = require("../../../database/utils/userRols");
 
 const getEstablishments = async (req, res, next) => {
   try {
@@ -53,18 +54,26 @@ const getEstablishments = async (req, res, next) => {
 };
 
 const deleteEstablishmentById = async (req, res, next) => {
-  const { idEstablishment } = req.params;
+  const { username, userRol } = req.user;
 
-  try {
-    await Establishment.findOneAndDelete({
-      idEstablishment,
-    });
+  if (userRol === rolAdmin) {
+    const { idEstablishment } = req.params;
+    try {
+      await Establishment.findOneAndDelete({
+        idEstablishment,
+      });
 
-    res.status(200).json({ msg: "The establishment has been deleted" });
-  } catch (error) {
-    error.statusCode = 400;
-    debug(chalk.red("Bad request"));
-    error.message = "Bad request";
+      res.status(200).json({ msg: "The establishment has been deleted" });
+    } catch (error) {
+      error.statusCode = 400;
+      debug(chalk.red("Bad request"));
+      error.message = "Bad request";
+      next(error);
+    }
+  } else {
+    const error = new Error("Only administrators can delete an establishment");
+    error.statusCode = 401;
+    debug(chalk.red(`${username} attempt to delete an establishment`));
     next(error);
   }
 };
